Going back to previous scoped approach with a single module handling it
