+++ libModularParser/src/main/java/redactedrice/modularparser/scope/ScopeSupporter.java
@@ -1,12 +1,39 @@
 package redactedrice.modularparser.scope;
 
-
 import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 
-public interface ScopeSupporter {
-    public void addScopeHandler(ScopeHandler parser);
+public interface ScopeSupporter {	
+    void addScopedModule(String module);
 
-    public boolean isVariableDefined(String var);
+    boolean handlesModule(String module);
 
-    public Map<String, Object> getAllVariables();
+    void pushScope(String scope);
+
+    void popScope();
+
+    void removeScope(String scope);
+
+    String currentScope();
+
+    // Splits the scope from the logical line if present or implicit is allowed. 
+    // Otherwise returns null
+    String[] splitScope(String logicalLine);
+
+    default boolean doesOwn(String module, Optional<String> scope, String name) {
+        return getOwner(scope, name) == module;
+    }
+    
+    String getOwner(Optional<String> scope, String name);
+
+    String getScope(String name);
+
+    Object getData(Optional<String> scope, String name, String owner);
+    
+    Set<String> getAllOwnedNames(Optional<String> scope, String owner);
+    
+    Map<String, Object> getAllOwnedData(Optional<String> scope, String owner);
+
+    boolean setData(String scope, String name, String owner, Object data);
 }
